Respond to Jim Bosch's feedback.
I've made Jobs arguments to run rather than task members, and
preferred Butler persistence instead of return values as the main way
for components to communicate. Discussion of migration to SuperTask
now assumes that work will remain in tasks as much as possible.

I've written up his suggestion for primarily extracting metrics as a
post-processing step (but stripped of a lot of SuperTask-specific
details).