Respond to Kian-Tat Lim's feedback.
My example for using Measurement.notes was fairly contrived because I
didn't have a good picture for how notes would be useful. When
discussion with Jonathan Sick suggested that notes are mainly for
recording implementation details of the measurement-making algorithm,
I decided it would be better to remove both the requirement and the
associated example.